DEV-3035 Fix all test compilation errors
