+++ .run/Tumor Reference COLO mini.run.xml
@@ -2,7 +2,7 @@
   <configuration default="false" name="Tumor Reference COLO mini" type="Application" factoryName="Application" singleton="false">
     <option name="MAIN_CLASS_NAME" value="com.hartwig.pipeline.PipelineMain" />
     <module name="cluster" />
-    <option name="PROGRAM_PARAMETERS" value="-profile development -set_id COLO829v003M -output_bucket pipeline-output-dev -sample_json sample_json/coloMini.json -run_germline_caller false -cloud_sdk $USER_HOME$/google-cloud-sdk/bin -run_id dataset -publish_db_load_event true -cleanup false" />
+    <option name="PROGRAM_PARAMETERS" value="-profile development -set_id COLO829v003M -output_bucket pipeline-output-dev -sample_json sample_json/coloMini.json -run_germline_caller false -cloud_sdk $USER_HOME$/google-cloud-sdk/bin -run_tag no-api -publish_db_load_event true -cleanup false -pubsub_environment pilot-1 -pubsub_workflow analysis" />
     <option name="VM_PARAMETERS" value="-Xmx32m" />
     <extension name="coverage">
       <pattern>
