+++ cluster/src/main/java/com/hartwig/pipeline/reruns/PersistedDataset.java
@@ -1,27 +1,10 @@
 package com.hartwig.pipeline.reruns;
 
-import static java.util.Optional.ofNullable;
-
-import java.util.Map;
 import java.util.Optional;
 
 import com.hartwig.pipeline.datatypes.DataType;
-import com.hartwig.pipeline.input.PipelineInput;
 import com.hartwig.pipeline.storage.GoogleStorageLocation;
 
-public class PersistedDataset {
-
-    private final Map<String, Map<String, Map<String, String>>> datasetMap;
-    private final String billingProject;
-
-    public PersistedDataset(final PipelineInput input, final String billingProject) {
-        this.billingProject = billingProject;
-        this.datasetMap = input.dataset();
-    }
-
-    public Optional<GoogleStorageLocation> path(final String sample, final DataType dataType) {
-        return ofNullable(datasetMap.get(dataType.toString().toLowerCase())).map(d -> d.get(sample))
-                .map(d -> d.get("path"))
-                .map(p -> GoogleStorageLocation.from(p, billingProject));
-    }
+public interface PersistedDataset {
+    Optional<GoogleStorageLocation> path(final String sample, final DataType dataType);
 }
