--- cluster/src/main/java/com/hartwig/pipeline/storage/SbpS3FileSource.java
@@ -1,11 +1,0 @@
-package com.hartwig.pipeline.storage;
-
-import java.util.function.Function;
-
-public class SbpS3FileSource implements Function<String, String> {
-
-    @Override
-    public String apply(final String file) {
-        return String.format("s3://%s", file);
-    }
-}
