+++ cluster/src/test/java/com/hartwig/pipeline/calling/sage/SageGermlineCallerTest.java
@@ -7,6 +7,7 @@ import java.util.List;
 import com.google.common.collect.ImmutableList;
 import com.hartwig.pipeline.input.SomaticRunMetadata;
 import com.hartwig.pipeline.stages.Stage;
+import com.hartwig.pipeline.stages.TestPersistedDataset;
 import com.hartwig.pipeline.storage.GoogleStorageLocation;
 import com.hartwig.pipeline.tertiary.TertiaryStageTest;
 import com.hartwig.pipeline.testsupport.TestInputs;
@@ -23,7 +24,7 @@ public class SageGermlineCallerTest extends TertiaryStageTest<SageOutput> {
     @Override
     protected Stage<SageOutput, SomaticRunMetadata> createVictim() {
         return new SageGermlineCaller(TestInputs.defaultPair(),
-                new NoopPersistedDataset(),
+                new TestPersistedDataset(),
                 TestInputs.REF_GENOME_37_RESOURCE_FILES);
     }
