+++ cluster/src/test/java/com/hartwig/pipeline/tertiary/purple/PurpleTest.java
@@ -14,6 +14,7 @@ import com.hartwig.pipeline.input.SomaticRunMetadata;
 import com.hartwig.pipeline.report.Folder;
 import com.hartwig.pipeline.resource.RefGenome37ResourceFiles;
 import com.hartwig.pipeline.stages.Stage;
+import com.hartwig.pipeline.stages.TestPersistedDataset;
 import com.hartwig.pipeline.storage.GoogleStorageLocation;
 import com.hartwig.pipeline.tertiary.TertiaryStageTest;
 import com.hartwig.pipeline.testsupport.TestInputs;
@@ -117,7 +118,7 @@ public class PurpleTest extends TertiaryStageTest<PurpleOutput> {
                 TestInputs.gripssSomaticProcessOutput(),
                 TestInputs.amberOutput(),
                 TestInputs.cobaltOutput(),
-                new NoopPersistedDataset(),
+                new TestPersistedDataset(),
                 Arguments.testDefaultsBuilder().shallow(true).build());
         assertThat(victim.tumorReferenceCommands(input()).get(0).asBash()).contains(
                 "-highly_diploid_percentage 0.88 -somatic_min_purity_spread 0.1");
